+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 0.0.2 (upcoming):
+    - `local.load_settings` now supports a list of configuration files.
     - Add missing pyyaml dependency.
     - Improve documentation of the modules.